Get things working, add a -s (scale) parameter
With this commit, it actually works. It finishes registering the mouse
device, since I was missing a bit that actually completes registration.
It also adds a -s scaling parameter, since the mouse was a bit too
zippy on the machine I was testing with. Finally, add a basic license
and readme, since this code is pure gold and everyone will want it.