Add support for exposing side touches
Add an option that exposes a virtual keyboard. This keyboard will send a
keydown event (whose keycode is determined on the commandline) whenever
there are touches to the right or left of the virtual trackpad (with Y
coordinates still in range of the trackpad). This represents palms which
may be resting to the side of the trackpad. Add a webpage which can
respond to a specific key (-k 85 via commandline), and hide some kittens
when that key is down.