Add overlay image cycling, sidekey at all Y positions
Use local images in the overlay so they can be easily replaced. Add
support to listen to to the tilde key which will cycle through up to 4
images.

Also, change the side key to key down if there are touches to the side
of the trackscreen at any vertical position, rather than aligned with
the trackscreen vertically. This should help with the flickering if you
put your whole hand down on the side.