Add left/right palm hiding independently
Research has asked for the kittens to be hidden independently for each
palm. This turned out to be not so easy, since the code wasn't tracking
coordinates for multiple touches.

Save the position of each finger, which we can then use to independently
determine whether the left side key, right side key, or both should be
held down. Also change overlay.html to listen to keys for left and right
independently.